Convert file to TS and change to use import type
Converted @modules/editor to ts
Switched to import type instead of import().editor.foo in @ui/customcss/editor
